requirements.txt
inotify
Pillow
flask
htmlmin
beautifulsoup4
python-slugify
python-frontmatter
markdown
jinja2